Additional refactoring
Move entirely away from lazy_static to once_cell, which doesn't require
macros. It otherwise has extremely similar behavior. As a bonus, it can
also be used on locals.

`guid()` is converted to a macro, as it's nothing more than a simple
function call.

In the feature filter, I've moved the parsing of features inside the
conditional, which allows for the string to remain heap allocated
throughout.

Use `pub(crate)` over `crate`, as it eliminates an otherwise unncessary
feature gate.

Requests in the `reddit` portion can be extracted via a macro, which
trivially expands to the appropriate code.

Finally, I've renamed some `struct`s to avoid repetition of the module
name, per a recommendation from clippy.