Even more refactoring
Derive getters, remove unnecessary `const_vec_new` feature

Use a macro for `ApiReturnType`, which masks the fact that we need
`Deserialize` - if Rust gains structural types, this can be trivially
changed.

I've also changed the `request` macro to use a simpler to read syntax;
it should be easily understandable to anyone, even those unfamiliar with
the macro.